socket 미들웨어 분리, 테스트
