+++ sprint-mission-5/src/websocket/emitters.ts
@@ -1,11 +1,10 @@
-import { PayloadForCommentNoti, PayloadForPriceNoti } from '../typings/notificationTypes';
 import { getIo } from './setupWebSocket';
-
-const io = getIo();
 
 export const emitNotification = <T>(userId: number, payload: T) => {
   try {
+    const io = getIo();
     io.to(`${userId}`).emit('notification', payload);
+    console.log('notification 전송 완료');
   } catch (error) {
     console.error(`Failed to emit notification for user ${userId}:`, error);
   }
